# docs/openapi.yaml
openapi: 3.0.1
info:
  title: 'EventXperience Backend API'
  description: 'Generated API reference for the EventXperience backend. Responses use the shared ApiResponse envelope unless noted otherwise.'
  version: v1
paths:
...
  /api/clubs:
    post:
      tags:
        - clubs
      operationId: Club_CreateClub
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                Name:
                  '$ref': '#/components/schemas/String14'
                Description:
                  '$ref': '#/components/schemas/String14'
                Clubtype:
                  '$ref': '#/components/schemas/String15'
                ClubImage:
                  '$ref': '#/components/schemas/IFormFile'
                Phone:
                  '$ref': '#/components/schemas/String15'
                Email:
                  '$ref': '#/components/schemas/String15'
        required: true
      responses:
        201:
          description: Created
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    get:
      tags:
        - clubs
      operationId: Club_GetClubs
      parameters:
        - name: search
          in: query
          schema:
            '$ref': '#/components/schemas/String'
        - name: clubType
          in: query
          schema:
            '$ref': '#/components/schemas/String'
        - name: sortBy
          in: query
          schema:
            type: integer
            default: 0
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
...
  '/api/clubs/{clubId}/posts':
    post:
      tags:
        - clubs
      operationId: ClubPost_CreatePost
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/ClubPostCreateRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/ClubPostCreateRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/ClubPostCreateRequest'
        required: true
      responses:
        201:
          description: Created
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubPostResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubPostResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubPostResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    get:
      tags:
        - clubs
      operationId: ClubPost_GetPosts
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
        - name: search
          in: query
          schema:
            '$ref': '#/components/schemas/String'
        - name: sortBy
          in: query
          schema:
            '$ref': '#/components/schemas/PostSortBy'
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubPostResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubPostResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubPostResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
  '/api/clubs/{clubId}/posts/{id}':
    put:
      tags:
        - clubs
      operationId: ClubPost_UpdatePost
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
        - name: id
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/ClubPostUpdateRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/ClubPostUpdateRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/ClubPostUpdateRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubPostResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubPostResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfClubPostResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    delete:
      tags:
        - clubs
      operationId: ClubPost_DeletePost
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
        - name: id
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int32'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  /api/admin/clubs/posts/reindex:
    post:
      tags:
        - admin
      operationId: AdminClubPost_ReindexPosts
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  /api/admin/clubs/posts:
    get:
      tags:
        - admin
      operationId: AdminClubPost_GetAllPosts
      parameters:
        - name: search
          in: query
          schema:
            '$ref': '#/components/schemas/String'
        - name: sortBy
          in: query
          schema:
            '$ref': '#/components/schemas/PostSortBy'
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubPostResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubPostResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfClubPostResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
...
components:
  schemas:
...
    ApiError:
      required:
        - code
      type: object
      properties:
        code:
          '$ref': '#/components/schemas/String'
        details:
          default: null
      nullable: true
...
    ApiResponseOfClubPostResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/ClubPostResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
    ApiResponseOfClubResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/ClubResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfObject:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
    ApiResponseOfPagedResponseOfClubPostResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/PagedResponseOfClubPostResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
    ApiResponseOfPagedResponseOfClubResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/PagedResponseOfClubResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    Boolean:
      type: boolean
...
    ClubPostCreateRequest:
      required:
        - title
        - content
      type: object
      properties:
        title:
          '$ref': '#/components/schemas/String17'
        content:
          '$ref': '#/components/schemas/String18'
        postType:
          '$ref': '#/components/schemas/PostType'
        isPinned:
          '$ref': '#/components/schemas/Boolean'
    ClubPostResponse:
      required:
        - id
        - clubId
        - userId
        - title
        - content
        - postType
        - likesCount
        - viewCount
        - isPinned
        - createdAt
        - updatedAt
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        clubId:
          '$ref': '#/components/schemas/Int32'
        userId:
          '$ref': '#/components/schemas/Int32'
        title:
          '$ref': '#/components/schemas/String'
        content:
          '$ref': '#/components/schemas/String'
        postType:
          '$ref': '#/components/schemas/PostType'
        likesCount:
          '$ref': '#/components/schemas/Int32'
        viewCount:
          '$ref': '#/components/schemas/Int32'
        isPinned:
          '$ref': '#/components/schemas/Boolean'
        createdAt:
          '$ref': '#/components/schemas/DateTime2'
        updatedAt:
          '$ref': '#/components/schemas/DateTime2'
      nullable: true
    ClubPostResponse2:
      required:
        - id
        - clubId
        - userId
        - title
        - content
        - postType
        - likesCount
        - viewCount
        - isPinned
        - createdAt
        - updatedAt
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        clubId:
          '$ref': '#/components/schemas/Int32'
        userId:
          '$ref': '#/components/schemas/Int32'
        title:
          '$ref': '#/components/schemas/String'
        content:
          '$ref': '#/components/schemas/String'
        postType:
          '$ref': '#/components/schemas/PostType'
        likesCount:
          '$ref': '#/components/schemas/Int32'
        viewCount:
          '$ref': '#/components/schemas/Int32'
        isPinned:
          '$ref': '#/components/schemas/Boolean'
        createdAt:
          '$ref': '#/components/schemas/DateTime2'
        updatedAt:
          '$ref': '#/components/schemas/DateTime2'
    ClubPostUpdateRequest:
      required:
        - title
        - content
      type: object
      properties:
        title:
          '$ref': '#/components/schemas/String17'
        content:
          '$ref': '#/components/schemas/String18'
        postType:
          '$ref': '#/components/schemas/PostType'
        isPinned:
          '$ref': '#/components/schemas/Boolean'
    ClubResponse:
      required:
        - id
        - name
        - description
        - clubtype
        - clubImage
        - memberCount
        - eventCount
        - avaliableEventCount
        - maxMemberCount
        - isPrivate
        - currentVersionNumber
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        ownerId:
          '$ref': '#/components/schemas/Int32'
        name:
          '$ref': '#/components/schemas/String'
        description:
          '$ref': '#/components/schemas/String'
        clubtype:
          '$ref': '#/components/schemas/String'
        clubImage:
          '$ref': '#/components/schemas/String'
        memberCount:
          '$ref': '#/components/schemas/Int32'
        eventCount:
          '$ref': '#/components/schemas/Int32'
        avaliableEventCount:
          '$ref': '#/components/schemas/Int32'
        maxMemberCount:
          '$ref': '#/components/schemas/Int32'
        isPrivate:
          '$ref': '#/components/schemas/Boolean'
        currentVersionNumber:
          '$ref': '#/components/schemas/Int32'
        isOwner:
          '$ref': '#/components/schemas/Boolean'
        isManager:
          '$ref': '#/components/schemas/Boolean'
        isVolunteer:
          '$ref': '#/components/schemas/Boolean'
        canManage:
          '$ref': '#/components/schemas/Boolean'
        phone:
          '$ref': '#/components/schemas/String2'
        email:
          '$ref': '#/components/schemas/String2'
        rating:
          '$ref': '#/components/schemas/Double'
        websiteUrl:
          '$ref': '#/components/schemas/String2'
        location:
          '$ref': '#/components/schemas/String2'
      nullable: true
    ClubResponse2:
      required:
        - id
        - name
        - description
        - clubtype
        - clubImage
        - memberCount
        - eventCount
        - avaliableEventCount
        - maxMemberCount
        - isPrivate
        - currentVersionNumber
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        ownerId:
          '$ref': '#/components/schemas/Int32'
        name:
          '$ref': '#/components/schemas/String'
        description:
          '$ref': '#/components/schemas/String'
        clubtype:
          '$ref': '#/components/schemas/String'
        clubImage:
          '$ref': '#/components/schemas/String'
        memberCount:
          '$ref': '#/components/schemas/Int32'
        eventCount:
          '$ref': '#/components/schemas/Int32'
        avaliableEventCount:
          '$ref': '#/components/schemas/Int32'
        maxMemberCount:
          '$ref': '#/components/schemas/Int32'
        isPrivate:
          '$ref': '#/components/schemas/Boolean'
        currentVersionNumber:
          '$ref': '#/components/schemas/Int32'
        isOwner:
          '$ref': '#/components/schemas/Boolean'
        isManager:
          '$ref': '#/components/schemas/Boolean'
        isVolunteer:
          '$ref': '#/components/schemas/Boolean'
        canManage:
          '$ref': '#/components/schemas/Boolean'
        phone:
          '$ref': '#/components/schemas/String2'
        email:
          '$ref': '#/components/schemas/String2'
        rating:
          '$ref': '#/components/schemas/Double'
        websiteUrl:
          '$ref': '#/components/schemas/String2'
        location:
          '$ref': '#/components/schemas/String2'
...
    DateTime2:
      type: string
      format: date-time
...
    Double:
      type: number
      format: double
      nullable: true
...
    IEnumerableOfClubPostResponse:
      type: array
      items:
        '$ref': '#/components/schemas/ClubPostResponse2'
    IEnumerableOfClubResponse:
      type: array
      items:
        '$ref': '#/components/schemas/ClubResponse2'
...
    IFormFile:
      type: string
      format: binary
    Int32:
      type: integer
      format: int32
...
    Int325:
      type: integer
      format: int32
      default: 1
    Int326:
      type: integer
      format: int32
      default: 20
...
    MessageResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    Object: {}
...
    PagedResponseOfClubPostResponse:
      required:
        - items
        - totalCount
        - page
        - pageSize
      type: object
      properties:
        items:
          '$ref': '#/components/schemas/IEnumerableOfClubPostResponse'
        totalCount:
          '$ref': '#/components/schemas/Int32'
        page:
          '$ref': '#/components/schemas/Int32'
        pageSize:
          '$ref': '#/components/schemas/Int32'
        totalPages:
          '$ref': '#/components/schemas/Int32'
      nullable: true
    PagedResponseOfClubResponse:
      required:
        - items
        - totalCount
        - page
        - pageSize
      type: object
      properties:
        items:
          '$ref': '#/components/schemas/IEnumerableOfClubResponse'
        totalCount:
          '$ref': '#/components/schemas/Int32'
        page:
          '$ref': '#/components/schemas/Int32'
        pageSize:
          '$ref': '#/components/schemas/Int32'
        totalPages:
          '$ref': '#/components/schemas/Int32'
      nullable: true
...
    PostSortBy:
      type: integer
      default: 0
    PostType:
      type: integer
...
    String:
      type: string
...
    String14:
      maxLength: 30
      minLength: 0
      type: string
    String15:
      maxLength: 30
      minLength: 0
      pattern: '^(sport|sports|academic|social|cultural|game|gaming|music|other)$'
      type: string
...
    String17:
      maxLength: 150
      minLength: 0
      type: string
    String18:
      maxLength: 2000
      minLength: 0
      type: string
...
    String2:
      type: string
      nullable: true
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Send the access token in the Authorization header as `Bearer {token}` for protected API routes.'
      scheme: bearer
      bearerFormat: JWT
...